PG-651: Changed Voice Actor List dialog to offer both OK and Cancel buttons when initially displayed via Cast Size Planning dialog. Clicking OK attempts to select Actual Cast Size option (if there are enough actors). Clicking Cancel leaves the pre-existing cast size option selected.
PG-653: Cast size on main form reflects actual number of groups in Project
PG-654: When user generates groups, the new groups are saved (replacing previously generated groups).
Also fixed wording (and the minimum cast size number shown) in the warning message when the number of actors is below the minimum.