WIP: Three failing tests
PG-841: Fixed problem when navigating to a block which is a continuation of a quote from a previous verse, so now it will not switch out of rainbow mode.
Also changed UI so that now the user selects which mode to work in, and Glyssen stays in that mode (but if they navigate to a book for which there is no reference text, that will pop them back into Select Character mode).