First pass at adjusting Russian reference text based on v. 099 of DG
Added option to ReferenceTextUtility to not attempt to add missing quotation marks to a text based on the ones in English.English.
Corrected a couple places in the English text where quotation marks were missing, as well as an errant delivery and a straight apostrophe.
Made ReferenceTextUtility correctly identify the location of the first difference we care about.