Moved Fonts.Default initialization into static constructor for TestProject so that all unit tests that rely on that can be run independently and not depend on the order of tests.
Added early return in MigrateProjectData to avoid logging the no-op case when we're already on the correct version. Note: This means that MigrateDeprecatedCharacterIds will NOT be called if the version is already up-to-date. (I think this should be okay.)
Prevented overwriting a "Needs Review" character assignment in MigrateDeprecatedCharacterIds. (There is some special logic in the parser to assign this when blocks include verses with both implicit and non-implicit characters, and the logic in this method was clobbering those.)
Fixed typos in comments in BlockNavigatorViewModelTests.cs and Block.cs