Fixed bugs whereby the current project is not remembered.
I couldn't see any fundamental change in the logic that would have caused this, but I could see that whenever a project was loaded on initial startup, the setting to save it was getting cleared. It was happening a second time AFTER setting the project completed, so that if you closed Glyssen without doing anything else. it wouldn't remember the project. There may be a timing issue involved, or it may just be that the font-related crash got me into an unusual scenario that revealed an existing problem.
Another issue was that following import, the modified date of the imported project was not getting set. While technically this could be considered correct, it made it hard to find the project in the listed (sorting by date) if there was a subsequent crash before the project got saved. This was an existing deficiency.