Attempt to prevent creation of release packages on nuget whenever a PR is merged into master
We only want releases to be created when a commit is tagged with a version number. All others shoud be beta.
This approach might not be correct since gitversion CLI returns a warning about ambiguity between the base gitversion.yml file (in the solution dir) and the ones in the project folders. But it appears that (at least on Appveyor) the yml file in the project folder is the one that wins.