Added logic to Block.TryMatchToReportingClause to limit matching to blocks that are narrator or "Needs Review"
Changed reftext matching logic to allow a "Needs Review" block to be treated as a match for the purpose of assigning a previously omitted reporting clause in the verse.
This is related to PG-1403 in some sense, since it addresses the confusion that is making it hard to reproduce the crash.